tweak: Add missing symbols
Workaround to enable compilation of sqlite and zlib. Eventually this
needs to be implemented in touchHLE but for now we can compile and hope
that for the most part missing symbols are not called.

Note: This is probably more ok with zlib as whats missing is only used in gz.